Write missing JavaDocs for new Metrics classes.
